web/requirements.txt
lxml==4.7.1
sqlalchemy==1.3.23
alembic==1.5.5
portalocker==2.2.1
psutil==5.8.0
mypy_extensions==0.4.3
thrift==0.13.0
gitpython==3.1.11
PyYAML==5.4.1
jira==3.2.0
python-dotenv==0.20.0
./api/py/codechecker_api/dist/codechecker_api.tar.gz
./api/py/codechecker_api_shared/dist/codechecker_api_shared.tar.gz
